Evaluate works on cpu + GCN model
Evaluate.py now works without a gpu. Added an implementation of the GCN model presented in Mao et al. (2019) Learning Trajectory Dependencies for Human Motion Prediction. The model takes in the same data as the DummyModel and does the DCT and IDCT transforms directly before and after the network. Check the paper for more info on the GCN and DCT.